Asset action confirmation modal
